Rename; Event names connected -> connect; Add jsdoc
