+++ examples/basic.js
@@ -1,4 +1,4 @@
-var Nut = require('../node-nut');
+var Nut = require('../index');
 
 nut = new Nut(3493, '10.30.21.11');
 
@@ -6,22 +6,22 @@ nut.on('error', err => {
     console.log('There was an error: ' + err);
 });
 
-nut.on('close', () => {
+nut.on('disconnect', () => {
     console.log('Connection closed.');
 });
 
-nut.on('ready', () => {
-    nut.GetUPSList((upslist, err) => {
+nut.on('connect', () => {
+    nut.getUpsList((upslist, err) => {
         if (err) console.log('Error: ' + err)
         console.log(upslist);
 
         let upsname = Object.keys(upslist)[0];
         
-        nut.GetUPSVars(upsname, (vars, err) => {
+        nut.getUpsVars(upsname, (vars, err) => {
             if (err) console.err('Error:', err);
             console.log(vars);
         });
     });
 });
 
-nut.start();
+nut.connect();
