+++ examples/promises.js
@@ -1,4 +1,4 @@
-var Nut = require('../node-nut');
+var Nut = require('../index');
 
 nut = new Nut(3493, '10.30.21.11');
 
@@ -6,14 +6,14 @@ nut.on('error', err => {
     console.log('There was an error: ' + err);
 });
 
-nut.on('close', () => {
+nut.on('disconnect', () => {
     console.log('Connection closed.');
 });
 
-nut.start().then(async () => {
-    let upslist = await nut.GetUPSList();
+nut.connect().then(async () => {
+    let upslist = await nut.getUpsList();
     let upsname = Object.keys(upslist)[0];
 
-    let vars = await nut.GetUPSVars(upsname);
+    let vars = await nut.getUpsVars(upsname);
     console.log(vars);
 });
